chore: add a more complex and larger input
The complex data is a reasonably sized JSON object.
The larger input is a load of Shakespeare plays - the
total size of all Shakespeare is about 5.2MB, so this
is a subset of around 1.3mb. The data itself isn't
important, just that we're putting a load of stuff
through Temporal and the encryption server.